codegen: relocate soroban builtins [auth - extendttl]
